drivers: platform: switch to linux-userspace platform
Introduce the linux-userspace platform, replacing the old linux
platform. The driver sources are moved from drivers/platform/linux/
to drivers/platform/linux-userspace/, and the C define is renamed
from LINUX_PLATFORM to LINUX_USERSPACE_PLATFORM to better reflect
what the platform actually is: a native Linux userspace application
communicating with hardware through kernel interfaces (spidev,
i2c-dev, gpio-cdev).
