+++ board_configs/linux-userspace/CMakePresets.json
@@ -1,0 +1,67 @@
+{
+	"version": 7,
+	"configurePresets": [
+		{
+			"name": "linux-userspace-base",
+			"hidden": true,
+			"description": "Base config for Linux userspace platform",
+			"generator": "Ninja"
+		},
+		{
+			"name": "native",
+			"description": "Native Linux host (no specific hardware)",
+			"inherits": "linux-userspace-base",
+			"cacheVariables": {
+				"BOARD": "native",
+				"PLATFORM": "linux-userspace",
+				"BOARD_CONFIG_FILE": "linux-userspace_defconfig"
+			}
+		},
+		{
+			"name": "native-i386",
+			"description": "Native Linux host, 32-bit (i386 multilib)",
+			"inherits": "linux-userspace-base",
+			"cacheVariables": {
+				"BOARD": "native-i386",
+				"PLATFORM": "linux-userspace",
+				"BOARD_CONFIG_FILE": "linux-userspace_defconfig",
+				"CMAKE_C_FLAGS": "-m32",
+				"CMAKE_CXX_FLAGS": "-m32",
+				"CMAKE_EXE_LINKER_FLAGS": "-m32"
+			}
+		},
+		{
+			"name": "rpi4",
+			"description": "Raspberry Pi 4 (aarch64 cross-compile)",
+			"inherits": "linux-userspace-base",
+			"cacheVariables": {
+				"BOARD": "rpi4",
+				"PLATFORM": "linux-userspace",
+				"BOARD_CONFIG_FILE": "linux-userspace_defconfig",
+				"CMAKE_C_COMPILER": "aarch64-linux-gnu-gcc"
+			}
+		},
+		{
+			"name": "rpi5",
+			"description": "Raspberry Pi 5 (aarch64 cross-compile)",
+			"inherits": "linux-userspace-base",
+			"cacheVariables": {
+				"BOARD": "rpi5",
+				"PLATFORM": "linux-userspace",
+				"BOARD_CONFIG_FILE": "linux-userspace_defconfig",
+				"CMAKE_C_COMPILER": "aarch64-linux-gnu-gcc"
+			}
+		},
+		{
+			"name": "zed-linux",
+			"description": "ZedBoard running Linux",
+			"inherits": "linux-userspace-base",
+			"cacheVariables": {
+				"BOARD": "zed-linux",
+				"PLATFORM": "linux-userspace",
+				"BOARD_CONFIG_FILE": "linux-userspace_defconfig",
+				"CMAKE_C_COMPILER": "arm-linux-gnueabihf-gcc"
+			}
+		}
+	]
+}
