+++ board_configs/linux-userspace/linux-userspace_defconfig
@@ -1,0 +1,8 @@
+CONFIG_PLATFORM_DRIVERS=y
+CONFIG_LINUX_USERSPACE=y
+
+CONFIG_GPIO_LINUX_USERSPACE=y
+CONFIG_I2C_LINUX_USERSPACE=y
+CONFIG_SPI_LINUX_USERSPACE=y
+CONFIG_UART_LINUX_USERSPACE=y
+CONFIG_TIMER_LINUX_USERSPACE=y
