+++ doc/sphinx/source/architecture.rst
@@ -58,7 +58,8 @@ backend:
 .. code-block::
 
    drivers/platform/
-   ├── aducm3029/   maxim/      pico/      stm32/     xilinx/  ...
+   ├── aducm3029/   maxim/      pico/      stm32/     xilinx/
+   ├── linux-userspace/  ...
 
 Each platform provides the vendor-SDK-backed implementation of the HAL: the
 code that actually configures a peripheral and moves bytes.
