+++ drivers/platform/linux-userspace/linux_gpio.h
@@ -1,10 +1,9 @@
 /*******************************************************************************
- *   @file   linux/linux_i2c.h
- *   @brief  Header containing extra types used by the I2C driver.
+ *   @file   linux_gpio.h
+ *   @brief  Header containing no_os_gpio_platform_ops used by the GPIO driver.
  *   @author Dragos Bogdan (dragos.bogdan@analog.com)
- *   @author Jamila Macagba (Jamila.Macagba@analog.com)
 ********************************************************************************
- * Copyright 2020, 2025(c) Analog Devices, Inc.
+ * Copyright 2020(c) Analog Devices, Inc.
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are met:
@@ -31,24 +30,12 @@
  * NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
  * EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 *******************************************************************************/
-#ifndef LINUX_I2C_H_
-#define LINUX_I2C_H_
-
-#include <stdint.h>
+#ifndef LINUX_GPIO_H_
+#define LINUX_GPIO_H_
 
 /**
- * @struct linux_i2c_init_param
- * @brief Structure holding the initialization parameters for Linux platform
- * specific I2C parameters.
+ * @brief Linux specific GPIO platform ops structure
  */
-struct linux_i2c_init_param {
-	/** I2C bus ID (/dev/i2c-"device_id") */
-	uint32_t device_id;
-};
+extern const struct no_os_gpio_platform_ops linux_gpio_ops;
 
-/**
- * @brief Linux specific I2C platform ops structure
- */
-extern const struct no_os_i2c_platform_ops linux_i2c_ops;
-
-#endif // LINUX_I2C_H_
+#endif // LINUX_GPIO_H_
