+++ drivers/platform/linux-userspace/linux_spi.h
@@ -1,6 +1,7 @@
 /*******************************************************************************
- *   @file   linux/linux_uart.h
- *   @brief  Header containing extra types used by the UART driver.
+ *   @file   linux_spi.h
+ *   @brief  Header containing extra types and spi_platform_ops used by the
+ *           SPI driver.
  *   @author Dragos Bogdan (dragos.bogdan@analog.com)
 ********************************************************************************
  * Copyright 2020(c) Analog Devices, Inc.
@@ -30,24 +31,12 @@
  * NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
  * EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 *******************************************************************************/
-#ifndef LINUX_UART_H_
-#define LINUX_UART_H_
-
-#include "no_os_uart.h"
+#ifndef LINUX_SPI_H_
+#define LINUX_SPI_H_
 
 /**
- * @struct linux_uart_init_param
- * @brief Structure holding the initialization parameters for Linux platform
- * specific UART parameters.
+ * @brief Linux specific SPI platform ops structure
  */
-struct linux_uart_init_param {
-	/** UART device ID (/dev/"device_id") */
-	const char *device_id;
-};
+extern const struct no_os_spi_platform_ops linux_spi_ops;
 
-/**
- * @brief Linux platform specific UART platform ops structure
- */
-extern const struct no_os_uart_platform_ops linux_uart_ops;
-
-#endif // LINUX_UART_H_
+#endif // LINUX_SPI_H_
