+++ iio/iio_app/iio_app.c
@@ -65,7 +65,7 @@
 #include "tcp_socket.h"
 #endif
 
-#ifdef LINUX_PLATFORM
+#ifdef LINUX_USERSPACE_PLATFORM
 #include "linux_socket.h"
 #include "tcp_socket.h"
 #endif
@@ -85,7 +85,7 @@ static inline uint32_t _calc_uart_xfer_time(uint32_t len, uint32_t baudrate)
 	return ms;
 }
 
-#if !defined(LINUX_PLATFORM) && !defined(NO_OS_NETWORKING) && !defined(NO_OS_LWIP_NETWORKING) && !defined(CONFIG_USB_UART_MAXIM) && !defined(CONFIG_USB_UART_STM32)
+#if !defined(LINUX_USERSPACE_PLATFORM) && !defined(NO_OS_NETWORKING) && !defined(NO_OS_LWIP_NETWORKING) && !defined(CONFIG_USB_UART_MAXIM) && !defined(CONFIG_USB_UART_STM32)
 static int32_t iio_print_uart_info_message(struct no_os_uart_desc **uart_desc,
 		struct no_os_uart_init_param *user_uart_params,
 		char *message, int32_t msglen)
@@ -109,7 +109,7 @@ static int32_t iio_print_uart_info_message(struct no_os_uart_desc **uart_desc,
 static int32_t print_uart_hello_message(struct no_os_uart_desc **uart_desc,
 					struct no_os_uart_init_param *user_uart_params)
 {
-#if defined(LINUX_PLATFORM) || defined(NO_OS_NETWORKING) || defined(NO_OS_LWIP_NETWORKING) || defined(CONFIG_USB_UART_MAXIM) || defined(CONFIG_USB_UART_STM32)
+#if defined(LINUX_USERSPACE_PLATFORM) || defined(NO_OS_NETWORKING) || defined(NO_OS_LWIP_NETWORKING) || defined(CONFIG_USB_UART_MAXIM) || defined(CONFIG_USB_UART_STM32)
 	return 0;
 #else
 	const char *uart_data_size[] = { "5", "6", "7", "8", "9" };
@@ -144,7 +144,7 @@ static int32_t print_uart_error_message(struct no_os_uart_desc **uart_desc,
 	uint32_t msglen = sprintf(message,
 				  "IIOD server failed with code %d.\n",
 				  (int)status);
-#if defined(LINUX_PLATFORM) || defined(NO_OS_NETWORKING) || defined(NO_OS_LWIP_NETWORKING) || defined(CONFIG_USB_UART_MAXIM) || defined(CONFIG_USB_UART_STM32)
+#if defined(LINUX_USERSPACE_PLATFORM) || defined(NO_OS_NETWORKING) || defined(NO_OS_LWIP_NETWORKING) || defined(CONFIG_USB_UART_MAXIM) || defined(CONFIG_USB_UART_STM32)
 	(void)msglen;
 	printf("%s", message);
 	return 0;
@@ -185,7 +185,7 @@ static int32_t lwip_network_setup(struct iio_app_desc *app,
 }
 #endif
 
-#if defined(NO_OS_NETWORKING) || defined(LINUX_PLATFORM)
+#if defined(NO_OS_NETWORKING) || defined(LINUX_USERSPACE_PLATFORM)
 static int32_t network_setup(struct iio_init_param *iio_init_param,
 			     struct no_os_uart_desc *uart_desc,
 			     void *irq_desc,
@@ -193,7 +193,7 @@ static int32_t network_setup(struct iio_init_param *iio_init_param,
 {
 	static struct tcp_socket_init_param socket_param;
 
-#ifdef LINUX_PLATFORM
+#ifdef LINUX_USERSPACE_PLATFORM
 	socket_param.net = &linux_net;
 #endif
 #ifdef ADUCM_PLATFORM
@@ -233,7 +233,7 @@ static int32_t network_setup(struct iio_init_param *iio_init_param,
 static int32_t uart_setup(struct no_os_uart_desc **uart_desc,
 			  struct no_os_uart_init_param *uart_init_par)
 {
-#if defined(LINUX_PLATFORM) || defined(NO_OS_LWIP_NETWORKING)
+#if defined(LINUX_USERSPACE_PLATFORM) || defined(NO_OS_LWIP_NETWORKING)
 	*uart_desc = NULL;
 	return 0;
 #endif
@@ -355,7 +355,7 @@ int iio_app_init(struct iio_app_desc **app,
 	status = lwip_network_setup(application, app_init_param, &iio_init_param);
 	if (status)
 		goto error;
-#elif defined(NO_OS_NETWORKING) || defined(LINUX_PLATFORM)
+#elif defined(NO_OS_NETWORKING) || defined(LINUX_USERSPACE_PLATFORM)
 	status = network_setup(&iio_init_param, uart_desc, application->irq_desc,
 			       app_init_param);
 	if (status < 0)
